requirements.txt
asgiref==3.3.4
Django==3.2
django-crispy-forms==1.11.2
django-filter==2.4.0
mysqlclient==2.0.3
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
xlrd==2.0.1
xlwt==1.3.0
